Move test_asynchronous.py to pytests
